Fixed center positioning problem

+++ roadCarnage/src/states/Game.java
@@ -7,8 +7,8 @@ public class Game extends StateBasedGame{
 
     public static final String GAME_NAME = "Road Carnage";
     public static final int GAMEPLAY_STATE = 0;
-    public static final int WIDTH = 1000;
-    public static final int HEIGHT = 700;
+    public static final float WIDTH = 1000;
+    public static final float HEIGHT = 700;
 
     public static void main(String[] args) {
         Game game = new Game(GAME_NAME);
